wip(#1102): extract emit/function.rs (compile_* + frame-slot helpers)
